requirements.txt
# Core dependencies
agno>=0.1.0
google-generativeai>=0.3.0
pinecone-client>=2.2.1
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.12.0

# API integrations
newsapi-python>=0.2.7
jina-reader>=0.1.0

# Optional dependencies
pydantic>=2.0.0
asyncio>=3.4.3
